+++ autogpt_platform/frontend/src/app/auth/confirm/route.ts
@@ -2,7 +2,7 @@ import { type EmailOtpType } from "@supabase/supabase-js";
 import { type NextRequest } from "next/server";
 
 import { redirect } from "next/navigation";
-import { createServerClient } from "@/lib/supabase/server";
+import getServerSupabase from "@/lib/supabase/getServerSupabase";
 
 // Email confirmation route
 export async function GET(request: NextRequest) {
@@ -12,7 +12,7 @@ export async function GET(request: NextRequest) {
   const next = searchParams.get("next") ?? "/";
 
   if (token_hash && type) {
-    const supabase = createServerClient();
+    const supabase = getServerSupabase();
 
     if (!supabase) {
       redirect("/error");
