+++ autogpt_platform/frontend/src/app/layout.tsx
@@ -10,7 +10,6 @@ import TallyPopupSimple from "@/components/TallyPopup";
 import { GoogleAnalytics } from "@next/third-parties/google";
 import { Toaster } from "@/components/ui/toaster";
 import { IconType } from "@/components/ui/icons";
-import { createServerClient } from "@/lib/supabase/server";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -24,17 +23,10 @@ export default async function RootLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const supabase = createServerClient();
-
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
   return (
     <html lang="en">
       <body className={cn("antialiased transition-colors", inter.className)}>
         <Providers
-          initialUser={user}
           attribute="class"
           defaultTheme="light"
           // Feel free to remove this line if you want to use the system theme by default
@@ -43,8 +35,6 @@ export default async function RootLayout({
         >
           <div className="flex min-h-screen flex-col items-center justify-center">
             <Navbar
-              user={user}
-              isLoggedIn={!!user}
               links={[
                 {
                   name: "Agent Store",
