+++ autogpt_platform/frontend/src/components/agptui/ProfileInfoForm.tsx
@@ -6,25 +6,17 @@ import { useState } from "react";
 import Image from "next/image";
 
 import { Button } from "./Button";
-
 import { IconPersonFill } from "@/components/ui/icons";
-
-import { AutoGPTServerAPI } from "@/lib/autogpt-server-api/client";
 import { CreatorDetails, ProfileDetails } from "@/lib/autogpt-server-api/types";
-import { createClient } from "@/lib/supabase/client";
 import { Separator } from "@/components/ui/separator";
+import useSupabase from "@/hooks/useSupabase";
+import { useBackendAPI } from "@/lib/autogpt-server-api/context";
 
 export const ProfileInfoForm = ({ profile }: { profile: CreatorDetails }) => {
   const [isSubmitting, setIsSubmitting] = useState(false);
   const [profileData, setProfileData] = useState(profile);
-
-  const supabase = createClient();
-
-  const api = new AutoGPTServerAPI(
-    process.env.NEXT_PUBLIC_AGPT_SERVER_URL,
-    process.env.NEXT_PUBLIC_AGPT_WS_SERVER_URL,
-    supabase,
-  );
+  const { supabase } = useSupabase();
+  const api = useBackendAPI();
 
   const submitForm = async () => {
     try {
