+++ autogpt_platform/frontend/src/components/agptui/RatingCard.tsx
@@ -3,8 +3,7 @@
 import * as React from "react";
 import { Cross1Icon } from "@radix-ui/react-icons";
 import { IconStar, IconStarFilled } from "@/components/ui/icons";
-import { createClient } from "@/lib/supabase/client";
-import { AutoGPTServerAPI } from "@/lib/autogpt-server-api/client";
+import { useBackendAPI } from "@/lib/autogpt-server-api/context";
 
 interface RatingCardProps {
   agentName: string;
@@ -18,18 +17,7 @@ export const RatingCard: React.FC<RatingCardProps> = ({
   const [rating, setRating] = React.useState<number>(0);
   const [hoveredRating, setHoveredRating] = React.useState<number>(0);
   const [isVisible, setIsVisible] = React.useState(true);
-
-  const supabase = React.useMemo(() => createClient(), []);
-
-  const api = React.useMemo(
-    () =>
-      new AutoGPTServerAPI(
-        process.env.NEXT_PUBLIC_AGPT_SERVER_URL,
-        process.env.NEXT_PUBLIC_AGPT_WS_SERVER_URL,
-        supabase,
-      ),
-    [supabase],
-  );
+  const api = useBackendAPI();
 
   const handleClose = () => {
     setIsVisible(false);
