+++ autogpt_platform/frontend/src/components/monitor/AgentFlowList.tsx
@@ -1,4 +1,4 @@
-import AutoGPTServerAPI, {
+import BackendAPI, {
   GraphExecution,
   GraphMeta,
 } from "@/lib/autogpt-server-api";
@@ -45,8 +45,6 @@ export const AgentFlowList = ({
   onSelectFlow: (f: GraphMeta) => void;
   className?: string;
 }) => {
-  const api = useMemo(() => new AutoGPTServerAPI(), []);
-
   return (
     <Card className={className}>
       <CardHeader className="flex-row items-center justify-between space-x-3 space-y-0">
