+++ autogpt_platform/frontend/src/lib/autogpt-server-api/index.ts
@@ -1,6 +1,6 @@
-import { AutoGPTServerAPI } from "./client";
+import BackendAPI from "./client";
 
-export default AutoGPTServerAPI;
+export default BackendAPI;
 export * from "./client";
 export * from "./types";
 export * from "./utils";
