--- autogpt_platform/frontend/src/lib/supabase/client.ts
@@ -1,13 +1,0 @@
-import { createBrowserClient } from "@supabase/ssr";
-
-export function createClient() {
-  try {
-    return createBrowserClient(
-      process.env.NEXT_PUBLIC_SUPABASE_URL!,
-      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-    );
-  } catch (error) {
-    console.error("error creating client", error);
-    return null;
-  }
-}
